refactor: move rule_required.py to lib/models
