+++ apps/users/views.py
@@ -18,11 +18,11 @@ from django.views.generic.base import TemplateView
 from apps.companies.models import CompanyFavorite
 from apps.jobs.models import Job, Job_Resume, JobFavorite
 from apps.resumes.models import Resume
+from lib.models.rule_required import rule_required
 
 from .forms import CustomUserCreationForm, UserInfoForm
 from .forms.users_form import PasswordResetForm
 from .models import UserInfo
-from .rule_required import rule_required
 
 
 def home(request):
